Apply black formatting

+++ dask/array/wrap.py
@@ -23,41 +23,50 @@ def _parse_wrap_args(func, args, kwargs, shape):
     if not isinstance(shape, (tuple, list)):
         shape = (shape,)
 
-    chunks = kwargs.pop('chunks', 'auto')
+    chunks = kwargs.pop("chunks", "auto")
 
-    dtype = kwargs.pop('dtype', None)
+    dtype = kwargs.pop("dtype", None)
     if dtype is None:
         dtype = func(shape, *args, **kwargs).dtype
     dtype = np.dtype(dtype)
 
     chunks = normalize_chunks(chunks, shape, dtype=dtype)
-    name = kwargs.pop('name', None)
+    name = kwargs.pop("name", None)
 
-    name = name or funcname(func) + '-' + tokenize(func, shape, chunks, dtype, args, kwargs)
+    name = name or funcname(func) + "-" + tokenize(
+        func, shape, chunks, dtype, args, kwargs
+    )
 
-    return {'shape': shape, 'dtype': dtype, 'kwargs': kwargs,
-            'chunks': chunks, 'name': name}
+    return {
+        "shape": shape,
+        "dtype": dtype,
+        "kwargs": kwargs,
+        "chunks": chunks,
+        "name": name,
+    }
 
 
 def wrap_func_shape_as_first_arg(func, *args, **kwargs):
     """
     Transform np creation function into blocked version
     """
-    if 'shape' not in kwargs:
+    if "shape" not in kwargs:
         shape, args = args[0], args[1:]
     else:
-        shape = kwargs.pop('shape')
+        shape = kwargs.pop("shape")
 
     if isinstance(shape, Array):
-        raise TypeError('Dask array input not supported. '
-                        'Please use tuple, list, or a 1D numpy array instead.')
+        raise TypeError(
+            "Dask array input not supported. "
+            "Please use tuple, list, or a 1D numpy array instead."
+        )
 
     parsed = _parse_wrap_args(func, args, kwargs, shape)
-    shape = parsed['shape']
-    dtype = parsed['dtype']
-    chunks = parsed['chunks']
-    name = parsed['name']
-    kwargs = parsed['kwargs']
+    shape = parsed["shape"]
+    dtype = parsed["dtype"]
+    chunks = parsed["chunks"]
+    name = parsed["name"]
+    kwargs = parsed["kwargs"]
 
     keys = product([name], *[range(len(bd)) for bd in chunks])
     shapes = product(*chunks)
@@ -74,21 +83,21 @@ def wrap_func_like(func, *args, **kwargs):
     """
     x = args[0]
     meta = meta_from_array(x)
-    shape = kwargs.get('shape', x.shape)
+    shape = kwargs.get("shape", x.shape)
 
     parsed = _parse_wrap_args(func, args, kwargs, shape)
-    shape = parsed['shape']
-    dtype = parsed['dtype']
-    chunks = parsed['chunks']
-    name = parsed['name']
-    kwargs = parsed['kwargs']
+    shape = parsed["shape"]
+    dtype = parsed["dtype"]
+    chunks = parsed["chunks"]
+    name = parsed["name"]
+    kwargs = parsed["kwargs"]
 
     keys = product([name], *[range(len(bd)) for bd in chunks])
     shapes = product(*chunks)
     shapes = list(shapes)
     kw = [kwargs for _ in shapes]
     for i, s in enumerate(list(shapes)):
-        kw[i]['shape'] = s
+        kw[i]["shape"] = s
     vals = ((partial(func, dtype=dtype, **k),) + args for (k, s) in zip(kw, shapes))
 
     dsk = dict(zip(keys, vals))
@@ -109,7 +118,7 @@ def wrap_func_like_safe(func, func_like, *args, **kwargs):
 
 @curry
 def wrap(wrap_func, func, **kwargs):
-    func_like = kwargs.pop('func_like', None)
+    func_like = kwargs.pop("func_like", None)
     if func_like is None:
         f = partial(wrap_func, func, **kwargs)
     else:
@@ -123,16 +132,16 @@ def wrap(wrap_func, func, **kwargs):
     Original signature follows below.
     """
     if func.__doc__ is not None:
-        f.__doc__ = template % {'name': func.__name__} + func.__doc__
-        f.__name__ = 'blocked_' + func.__name__
+        f.__doc__ = template % {"name": func.__name__} + func.__doc__
+        f.__name__ = "blocked_" + func.__name__
     return f
 
 
 w = wrap(wrap_func_shape_as_first_arg)
 
-ones = w(np.ones, dtype='f8')
-zeros = w(np.zeros, dtype='f8')
-empty = w(np.empty, dtype='f8')
+ones = w(np.ones, dtype="f8")
+zeros = w(np.zeros, dtype="f8")
+empty = w(np.empty, dtype="f8")
 full = w(np.full)
